Adds Juju release publish github action
This action will be triggered whenever a change is detected in the releases/
folder.

If there is a new release that was not yet published on charmhub, it will upload
the respective FINOS Legend Engine, SDLC, and Studio images into charmhub and
make a release. It will also create a new release on latest.